+++ examples/data-objects/primitives/src/model.tsx
